requirements.txt
alembic==1.7.5
attrs==21.4.0
autopep8==1.6.0
click==8.0.3
faker==25.0.0
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
pytest-assume==2.4.3
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.2
zipp==3.7.0
